training_code/requirements.txt
torch
numpy
Pillow
opencv-python
tqdm
tensorboard
pyyaml
safetensors